+++ src/models/publication/kind/post.rs
@@ -4,7 +4,7 @@ use crate::models::publication::{PublicationInheritor, PublicationKind};
 #[derive(Default, Clone, Debug)]
 pub struct Post {
     /// An identifier of the corresponding publication. Should always be set to a valid value if
-    /// constructing with `{ ... }` and match [Publication::id][crate::models::Publication::id]
+    /// constructing with `{ ... }` and match [`Publication::id`][crate::models::Publication::id]
     pub id: u64,
     // /// The post's content
     // pub pages: Vec<Page>,
@@ -15,7 +15,7 @@ pub struct Post {
     /// The post's rubric name
     pub rubric_name: Option<String>,
     /// The post's rubric karma coefficient
-    pub rubric_karma_coef: Option<f32>,
+    pub rubric_karma_coef: Option<f64>,
     // /// The post's relay race
     // pub relay_race: Option<RelayRace>,
 }
